# backend/v2/openapi.yaml
openapi: 3.0.3
info:
  title: AI Skill Management Platform
  version: 2.0.0
  description: Multi-tenant SaaS for creating, managing, assigning, and executing AI skills
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/skills:
    get:
      summary: List skills (paginated, filtered)
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, minimum: 1, maximum: 100 } }
        - { name: search, in: query, schema: { type: string, default: "" } }
        - { name: category, in: query, schema: { type: string, default: "" } }
        - { name: status, in: query, schema: { type: string, default: "" } }
      responses:
        '200':
          description: Paginated skills list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create a new skill
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 255 }
                description: { type: string, maxLength: 5000 }
                content: { type: string, maxLength: 50000 }
                category: { type: string, default: general }
      responses:
        '201':
          description: Skill created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/skills/{skill_id}:
    get:
      summary: Get skill by ID
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: skill_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Skill detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Full update skill (admin)
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: skill_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                content: { type: string }
                category: { type: string }
      responses:
        '200':
          description: Skill updated
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      summary: Partial update skill (admin)
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: skill_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                content: { type: string }
                category: { type: string }
      responses:
        '200':
          description: Skill patched
    delete:
      summary: Soft delete skill (admin)
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: skill_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Skill deleted
...
  /api/skills/{skill_id}/assignments:
    get:
      summary: Get who this skill is assigned to
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: skill_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Assignment list
...
  /api/users/{user_id}/skills:
    get:
      summary: Get skills assigned to user
      tags: [Users]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: User skills list
    post:
      summary: Assign skills to user (admin)
      tags: [Users]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill_ids]
              properties:
                skill_ids: { type: array, items: { type: string, format: uuid } }
      responses:
        '201':
          description: Skills assigned
  /api/users/{user_id}/skills/{skill_id}:
    delete:
      summary: Remove skill from user (admin)
      tags: [Users]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: skill_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Skill removed
...
  /api/teams/{team_id}/skills:
    get:
      summary: Get skills assigned to team
      tags: [Teams]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: team_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Team skills list
    post:
      summary: Assign skills to team (admin)
      tags: [Teams]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: team_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill_ids]
              properties:
                skill_ids: { type: array, items: { type: string, format: uuid } }
      responses:
        '201':
          description: Skills assigned to team
...
  /api/assignments:
    post:
      summary: Bulk assign skills to users/teams
      tags: [Assignments]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill_id, assignee_type, assignee_id]
              properties:
                skill_id: { type: string, format: uuid }
                assignee_type: { type: string, enum: [user, team, org] }
                assignee_id: { type: string, format: uuid }
                expires_at: { type: string, format: date-time }
      responses:
        '201':
          description: Assignment created
    delete:
      summary: Bulk unassign skills
      tags: [Assignments]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill_id, assignee_type, assignee_id]
              properties:
                skill_id: { type: string, format: uuid }
                assignee_type: { type: string }
                assignee_id: { type: string, format: uuid }
      responses:
        '200':
          description: Assignments removed
    get:
      summary: List all assignments with filters (paginated)
      tags: [Assignments]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20 } }
        - { name: skill_id, in: query, schema: { type: string, format: uuid } }
        - { name: assignee_type, in: query, schema: { type: string } }
        - { name: assignee_id, in: query, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Paginated assignments list
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: AUTHORIZATION_ERROR
              message: Admin role required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Skill not found
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: CONFLICT
              message: A skill with this name already exists
  schemas:
    SuccessResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data: { type: object }
        meta:
          type: object
          properties:
            page: { type: integer }
            total: { type: integer }
            per_page: { type: integer }
            total_pages: { type: integer }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: array, items: { type: string } }
    PaginatedResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { type: array, items: { type: object } }
        meta:
          type: object
          properties:
            page: { type: integer }
            total: { type: integer }
            per_page: { type: integer }
            total_pages: { type: integer }